@@ -7,18 +7,17 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
+    "@vue/compiler-sfc": "^3.5.0",
     "axios": "^1.7.3",
     "core-js": "^3.8.3",
     "vault-project": "file:",
-    "vue": "^2.7.16",
-    "vue-router": "^3.5.1",
-    "vuex": "^3.6.2"
+    "vue": "^3.5.0",
+    "vue-router": "^4.6.3",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~5.0.0",
     "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-plugin-vuex": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "vue-template-compiler": "^2.6.14"
+    "@vue/cli-plugin-vuex": "~5.0.0"
   }
 }
